+++ packages/dsw-mailer/CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed error reporting on worker startup
+
 
 ## [4.14.0]